Optimizations:
HexagonalWeightedGrid: created a cache with the distances between each
node to avoid recalculate them.

HexagonalProximityBuilder: reduced the number of operations to calculate
the total distances in a grid.
The algorithm, in order to achieve fast a good approximation, does the
assumption that the the most used keys should be closer to the center,
so it places first the most used key in the center, the following 6 most
used keys around it, the following 12 most used key around it, and
successively.
To analyze all the permutations of k keys in a radius, the algorithm
places one key and recursively create all the permutations of the other
keys: to analyze a permutation, it rotates the outer keys around the
inner keys. This produces all the permutations and gives the chance to
do some optimizations.
The implemented optimization consists in calculating once the distance
of each key in each position of the outer nodes to inner nodes, and
given that the distances between the outer nodes is constant, to
calculate the better distribution of the rotations of one permutation it
is necessary calculate one time the distances of the outer radius and
sum the distances to the inner radius.

With these optimizations the algorithm is two times faster.